+++ src/main/scala/com/databricks/labs/mosaic/expressions/raster/RST_FromBands.scala
@@ -41,7 +41,16 @@ case class RST_FromBands(
       *   The stacked and resampled tile.
       */
     override def rasterTransform(rasters: Seq[RasterTile]): Any = {
-        rasters.head.copy(raster = MergeBands.merge(rasters.map(_.raster), "bilinear", Option(exprConfig)))
+        //scalastyle:off println
+        //println(s"RST_FromBands - rasters length? ${rasters.length}")
+        //println(s"RST_FromBands - head createInfo: ${rasters.head.raster.getCreateInfo}")
+        //println(s"RST_FromBands - is head empty? ${rasters.head.raster.isEmptyRasterGDAL}")
+        //scalastyle:on println
+        rasters.head.copy(
+            raster = MergeBands.merge(
+                rasters.map(_.raster), "bilinear", Option(exprConfig)
+            )
+        )
     }
 
 }