+++ src/test/scala/org/apache/spark/sql/test/SharedSparkSessionGDAL.scala
@@ -45,11 +45,10 @@ trait SharedSparkSessionGDAL extends SharedSparkSession {
         sc.conf.set(MOSAIC_GDAL_NATIVE, "true")
         sc.conf.set(MOSAIC_TEST_MODE, "true")
         sc.conf.set(MOSAIC_MANUAL_CLEANUP_MODE, "false")
-        sc.conf.set(MOSAIC_CLEANUP_AGE_LIMIT_MINUTES, "15") // manual is -1 (default is 30)
-        sc.conf.set(MOSAIC_RASTER_USE_CHECKPOINT, MOSAIC_RASTER_USE_CHECKPOINT_DEFAULT)
+        sc.conf.set(MOSAIC_CLEANUP_AGE_LIMIT_MINUTES, "5") // manual is -1 (default is 30)
+        sc.conf.set(MOSAIC_RASTER_USE_CHECKPOINT, "true") // default is "false"
         sc.conf.set(MOSAIC_RASTER_CHECKPOINT, mosaicCheckpointRootDir)
         sc.conf.set(MOSAIC_RASTER_TMP_PREFIX, MOSAIC_RASTER_TMP_PREFIX_DEFAULT)
-        sc.conf.set(MOSAIC_RASTER_USE_CHECKPOINT, MOSAIC_RASTER_USE_CHECKPOINT_DEFAULT)
 
         Try(MosaicGDAL.enableGDAL(sc))
         Try(gdal.AllRegister())